% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{esr_rss_2r}
\alias{esr_rss_2r}
\title{Extract Results from rss_2r Simulation}
\format{
\subsection{\code{esr_rss_2r}}{

Large list (4385 elements)
}
}
\source{
#private
}
\usage{
esr_rss_2r
}
\description{
Sample Output from extract_simulation_results on rss_r2 data set.
}
\keyword{datasets}
