% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rss_2r}
\alias{rss_2r}
\title{Run Simulation Study 2 Reps}
\format{
\subsection{\code{rss_2r}}{

Large list (2 elements)
}
}
\source{
#private
}
\usage{
rss_2r
}
\description{
Sample Output from run_simulation_study with n_reps = 2.
}
\keyword{datasets}
