% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{run_cbc_models}
\alias{run_cbc_models}
\title{Run Case-by-case (cbc) Regression Models}
\usage{
run_cbc_models(sim_data, stop_zeroSD = FALSE, level = 0.95)
}
\arguments{
\item{sim_data}{A data frame of simulated data, created by \code{LCGA_GMM_sim}.
It should include columns for the time point and response variable.}

\item{stop_zeroSD}{A logical value indicating whether to stop (throw an error) when
the estimated standard deviation is zero. Default is FALSE, which means a warning is issued
and NULL is returned if the standard deviation is zero.}

\item{level}{Confidence level for the bootstrap confidence intervals. Default is 0.95.}
}
\value{
A list of model results. The list includes one element: 'ols',
which stores the result of the ordinary least squares regression model.
}
\description{
This function runs case-by-case (cbc) regression models on simulated data
using the \code{cbc_lm} function from the \code{OLStrajr} package. It includes a built-in
error handling mechanism to manage potential errors during model fitting.
}
\examples{
sim_data <- LCGA_GMM_sim(n_pers = 100, n_time = 5, beta_int = 0, beta_slo_time = 1,
mean_i = 0, var_i = 1, mean_s = 0, var_s = 1, cov_is = 0, mean_r = 0, var_r = 1)
model_results <- run_cbc_models(sim_data)

}
\references{
Carrig, M. M., Wirth, R. J., & Curran, P. J. (2004). A SAS Macro for Estimating and Visualizing
Individual Growth Curves. Structural Equation Modeling: A Multidisciplinary Journal, 11(1), 132-149.
doi:10.1207/S15328007SEM1101_9
}
