% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{run_lcga_models}
\alias{run_lcga_models}
\title{Run Latent Class Growth Analysis (LCGA) Models}
\usage{
run_lcga_models(sim_data, ng = 1)
}
\arguments{
\item{sim_data}{A data frame of simulated data, created by \code{LCGA_GMM_sim}.
It should include columns for the individual ID, time point, and response variable.}

\item{ng}{Number of classes/groups to be used in the LCGA models. Default is 1.}
}
\value{
A list of model results. The list includes two elements: 'ri',
which stores the result of the random intercept model,
and 'ris', which stores the result of the random intercept and slope model.
}
\description{
This function runs Latent Class Growth Analysis (LCGA) models on simulated data.
It uses the \code{hlme} function from the \code{lcmm} package to fit the models.
Two models are fitted: a random intercept model, and a random intercept and slope model.
}
\examples{
sim_data <- LCGA_GMM_sim(n_pers = 100, n_time = 5, beta_int = 0, beta_slo_time = 1,
mean_i = 0, var_i = 1, mean_s = 0, var_s = 1, cov_is = 0, mean_r = 0, var_r = 1)
model_results <- run_lcga_models(sim_data, ng = 1)

}
\references{
Proust-Lima, C., Philipps, V., & Liquet, B. (2017). Estimation of Extended Mixed Models Using
Latent Classes and Latent Processes: The R Package lcmm. Journal of Statistical Software, 78(2),
1-56. doi:10.18637/jss.v078.i02
}
