% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{run_lm_models}
\alias{run_lm_models}
\title{Run Ordinary Least Squares (OLS) Regression Model}
\usage{
run_lm_models(sim_data)
}
\arguments{
\item{sim_data}{A data frame of simulated data, created by \code{LCGA_GMM_sim}.
It should include columns for the time point and response variable.}
}
\value{
A list of model results. The list includes one element: 'ols',
which stores the result of the ordinary least squares regression model.
}
\description{
This function runs an Ordinary Least Squares (OLS) regression model on simulated data
using the \code{lm} function from the \code{stats} package.
}
\examples{
sim_data <- LCGA_GMM_sim(n_pers = 100, n_time = 5, beta_int = 0, beta_slo_time = 1,
mean_i = 0, var_i = 1, mean_s = 0, var_s = 1, cov_is = 0, mean_r = 0, var_r = 1)
model_results <- run_lm_models(sim_data)

}
\references{
R Core Team (2021). R: A language and environment for statistical computing. R Foundation for
Statistical Computing, Vienna, Austria. URL https://www.R-project.org/.
}
