% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{run_lme4_models}
\alias{run_lme4_models}
\title{Run Linear Mixed-Effects Models}
\usage{
run_lme4_models(sim_data)
}
\arguments{
\item{sim_data}{A data frame of simulated data, created by \code{LCGA_GMM_sim}.
It should include columns for the individual ID, time point, and response variable.}
}
\value{
A list of model results. The list includes two elements: 'ri', which stores
the result of the random intercept model, and 'ris', which stores the result of the
random intercept and slope model.
}
\description{
This function runs linear mixed-effects models on simulated data using the \code{lmer}
function from the \code{lme4} package. Two models are fitted: a random intercept model,
and a random intercept and slope model.
}
\examples{
sim_data <- LCGA_GMM_sim(n_pers = 100, n_time = 5, beta_int = 0, beta_slo_time = 1,
mean_i = 0, var_i = 1, mean_s = 0, var_s = 1, cov_is = 0, mean_r = 0, var_r = 1)
model_results <- run_lme4_models(sim_data)

}
\references{
Bates D, Maechler M, Bolker B, Walker S (2015). "Fitting Linear Mixed-Effects Models Using lme4."
Journal of Statistical Software, 67(1), 1-48. doi: 10.18637/jss.v067.i01.
}
