% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_crossing.R
\name{scenario_crossing}
\alias{scenario_crossing}
\title{Generate Simulated Data Across Various Scenarios}
\usage{
scenario_crossing(
  n_pers = c(50, 100),
  n_time = c(3, 4, 5),
  beta_int = c(0),
  beta_slo_time = c(0.1, 0.3, 0.5),
  beta_slo_covar = c(0.25, 0.5),
  beta_slo_interact = c(0, 1.5),
  mean_i = c(5, 10),
  var_i = c(0.05, 0.13),
  mean_s = c(0),
  var_s = c(0.05, 0.09),
  cov_is = c(0),
  mean_r = c(0),
  var_r = c(1, 1.2)
)
}
\arguments{
\item{n_pers}{A vector of number of individuals in the simulation. Default is
c(50, 100).}

\item{n_time}{A vector of number of time points in the simulation. Default is
c(3,4,5).}

\item{beta_int}{A vector of fixed intercept(s) for the linear model. Default
is 0.}

\item{beta_slo_time}{A vector of fixed effect(s) of time (slope) for the linear
model. Default is c(0.1, 0.3, 0.5).}

\item{beta_slo_covar}{A vector of fixed effect(s) of the covariate for the linear
model. Default is c(0.25, 0.5).}

\item{beta_slo_interact}{A vector of fixed effect(s) of the time-covariate interaction
for the linear model. Default is c(0, 1.5).}

\item{mean_i}{A vector of mean(s) of the random intercepts. Default is c(5, 10).}

\item{var_i}{A vector of variance(s) of the random intercepts. Default is
c(0.05, 0.13).}

\item{mean_s}{A vector of mean(s) of the random slopes. Default is 0.}

\item{var_s}{A vector of variance(s) of the random slopes. Default is
c(0.05, 0.09).}

\item{cov_is}{A vector of covariance(s) between the random intercepts and slopes.
Default is 0.}

\item{mean_r}{A vector of mean(s) of the Gaussian noise added to the response
variable. Default is 0.}

\item{var_r}{A vector of variance(s) of the Gaussian noise added to the response
variable. Default is c(1 , 1.2).}
}
\value{
A list of data frames, each representing a specific scenario. Each data
frame includes simulated longitudinal data
generated by the \code{LCGA_GMM_sim} function.
}
\description{
This function generates simulated longitudinal data across various scenarios
by calling \code{LCGA_GMM_sim}. It creates a list of data frames, each representing
a specific scenario specified by combinations of the function parameters.
}
\examples{
# Generate a list of data frames representing different scenarios
scenarios <- scenario_crossing(n_pers=c(100, 200), n_time=c(5, 6))

}
\seealso{
\code{\link{LCGA_GMM_sim}}
}
