% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ssr_rss_2r}
\alias{ssr_rss_2r}
\title{Summarize Simulation Results from rss_2r}
\format{
\subsection{\code{ssr_rss_2r}}{

Data frame (2281 obs. & 32 variables)
}
}
\source{
#private
}
\usage{
ssr_rss_2r
}
\description{
Sample Output from summarize_simulation_results on esr_rss_2r data set.
}
\keyword{datasets}
